os-terraform: Use ssh keypair in misc-files
Generating a kepair is unnecessary. Use misc/files/id_shared and
misc/files/id_shared.pub as key files for the cluster. This also makes
the Terraform environment use the same keys as the Heat environment.
Environment similarity is good.

Make a symlink from ssh/id_caasp to the misc/files/id_shared file so
that existing tooling that may use this file will not break.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>