os-terraform: Support ECP and OVH networking
Networking needed a lot of work to support both OVH and ECP.

For ECP, Terraform can now optionally create an internal net and will
create a router to the external net for it.

For OVH, floating IPs are not supported, nor is creation of internal
networks, so floating IP creation is now done optionally.

Sections started running together, so I made headings.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>